stream: split stream_impl.c and add stream_enqueue.c
The implementation of enqueue functions are quite distinct. Move them to
separate soruce file for better organization.
