+++ src/mpi/stream/Makefile.mk
@@ -4,4 +4,5 @@
 ##
 
 mpi_core_sources +=   \
-    src/mpi/stream/stream_impl.c
+    src/mpi/stream/stream_impl.c \
+    src/mpi/stream/stream_enqueue.c
